Refactor functions into CartService
